implementing policy and permissions

+++ app/Http/Controllers/PostController.php
@@ -10,6 +10,7 @@ use Illuminate\Auth\Events\Validated;
 class PostController extends Controller
 {
     public function index(){
+        $this->authorize('viewAny', Post::class);
 
         $orderColumn = request('order_column','created_at');
         if(!in_array($orderColumn,['id','title','created_at'])){
@@ -38,10 +39,12 @@ class PostController extends Controller
     }
 
     public function create(){
+        $this->authorize('create', Post::class);
         return inertia('Posts/Create');
     }
 
     public function store(StorePostRequest $request){
+        $this->authorize('create', Post::class);
         Post::create($request->validated());
 
         return redirect()->route('posts.index')
@@ -49,16 +52,19 @@ class PostController extends Controller
     }
 
     public function destroy(Post $post){
+        $this->authorize('create', Post::class);
         $post->delete();
          return redirect()->route('posts.index')
             ->with('message', 'Post deleted succesfully.');
     }
 
     public function edit(Post $post){
+        $this->authorize('create', Post::class);
         return inertia('Posts/Edit',compact('post'));
     }
 
     public function update(Post $post, StorePostRequest $request){
+        $this->authorize('create', Post::class);
         $post->update($request->validated());
         return redirect()->route('posts.index')
             ->with('message', 'Post updated ');
